Address PR review comments for vector Parquet read/write support
- Use VectorLogicalType.VECTOR_BYTE_ORDER instead of hardcoded
  ByteOrder.LITTLE_ENDIAN in all 4 locations (write support, reader,
  Scala reader context, file group format)
- Add Math.multiplyExact overflow guard for dimension * elementSize
  in HoodieRowParquetWriteSupport
- Remove unnecessary array clone in HoodieSparkParquetReader
- Add clarifying comment on non-vector column else branch
- Fix misleading "float arrays" comment to "typed arrays"
- Move inline JavaConverters import to top-level in
  SparkFileFormatInternalRowReaderContext
- Import Metadata at top level instead of fully-qualified reference
- Consolidate duplicate detectVectorColumns, replaceVectorColumnsWithBinary,
  and convertBinaryToVectorArray into SparkFileFormatInternalRowReaderContext
  companion object; HoodieFileGroupReaderBasedFileFormat now delegates
- Add Javadoc on VectorType explaining it's needed for InternalSchema
  type hierarchy (cannot reuse HoodieSchema.Vector)
- Clean up unused imports (ByteOrder, ByteBuffer, GenericArrayData,
  StructField, BinaryType, HoodieSchemaType)
